+++ src/dune/cinaps.ml
@@ -81,13 +81,14 @@ let gen_rules sctx t ~dir ~scope =
     Modules.exe_unwrapped modules
     |> Modules.map_user_written ~f:(Preprocessing.pp_module preprocess)
   in
+  let dune_version = Scope.project scope |> Dune_project.dune_version in
   let compile_info =
     Lib.DB.resolve_user_written_deps_for_exes (Scope.libs scope)
       [ (t.loc, name) ]
       ( Lib_dep.Direct (loc, Lib_name.of_string_exn "cinaps.runtime" ~loc:None)
       :: t.libraries )
       ~pps:(Dune_file.Preprocess_map.pps t.preprocess)
-      ~variants:None ~optional:false
+      ~dune_version ~variants:None ~optional:false
   in
   let cctx =
     Compilation_context.create () ~super_context:sctx ~expander ~scope ~obj_dir