+++ src/dune/opam_create.ml
@@ -45,7 +45,7 @@ let package_fields
     ; name = _
     ; path = _
     ; version = _
-    ; kind = _
+    ; has_opam_file = _
     ; tags
     ; loc = _
     ; deprecated_package_names = _
@@ -105,9 +105,7 @@ let opam_fields project (package : Package.t) =
   in
   let package_fields = package_fields package ~project in
   let open Opam_file.Create in
-  let info =
-    Package.Info.superpose (Dune_project.info project) package.Package.info
-  in
+  let info = package.Package.info in
   let optional_fields =
     [ ("bug-reports", Package.Info.bug_reports info)
     ; ("homepage", Package.Info.homepage info)
@@ -115,9 +113,8 @@ let opam_fields project (package : Package.t) =
     ; ("license", Package.Info.license info)
     ; ("version", Dune_project.version project)
     ; ( "dev-repo"
-      , Option.map
-          ~f:(Format.asprintf "%a" Package.Source_kind.pp)
-          (Package.Info.source info) )
+      , Option.map ~f:Package.Source_kind.to_string (Package.Info.source info)
+      )
     ]
     |> List.filter_map ~f:(fun (k, v) ->
            Option.map v ~f:(fun v -> (k, string v)))
@@ -195,6 +192,4 @@ let add_rules sctx ~dir =
   if Dune_project.generate_opam_files project then
     Dune_project.packages project
     |> Package.Name.Map.iter ~f:(fun (pkg : Package.t) ->
-           match pkg.kind with
-           | Dune _ -> add_rule sctx ~project ~pkg
-           | Opam -> ())
+           add_rule sctx ~project ~pkg)