+++ src/dune/package.mli
@@ -21,7 +21,7 @@ module Name : sig
 
   val to_dyn : t -> Dyn.t
 
-  val of_basename : string -> t option
+  val of_opam_file_basename : string -> t option
 end
 
 module Dependency : sig
@@ -62,14 +62,6 @@ module Dependency : sig
   val decode : t Dune_lang.Decoder.t
 end
 
-module Kind : sig
-  type has_opam = bool
-
-  type t =
-    | Dune of has_opam
-    | Opam
-end
-
 module Source_kind : sig
   type t =
     | Github of string * string
@@ -77,7 +69,7 @@ module Source_kind : sig
 
   val to_dyn : t Dyn.Encoder.t
 
-  val pp : Format.formatter -> t -> unit
+  val to_string : t -> string
 
   val decode : t Dune_lang.Decoder.t
 end
@@ -122,7 +114,7 @@ type t =
   ; info : Info.t
   ; path : Path.Source.t
   ; version : string option
-  ; kind : Kind.t
+  ; has_opam_file : bool
   ; tags : string list
   ; deprecated_package_names : Loc.t Name.Map.t
   }
@@ -142,3 +134,6 @@ val to_dyn : t -> Dyn.t
 val hash : t -> int
 
 val is_opam_file : Path.t -> bool
+
+(** Construct a package description from an opam file. *)
+val load_opam_file : Path.Source.t -> Name.t -> t